requirements.txt
tortoise-orm
numpy
loguru
httpx
aiofiles